web_assistant/requirements.txt
Flask==3.1.2
openai==1.102.0
google-api-python-client==2.179.0
google-auth==2.40.3
python-dotenv==1.0.1
pytest==8.4.1
